+++ internal/payments/models.go
@@ -1,7 +1,7 @@
 package payments
 
 type PostPaymentRequest struct {
-	CardNumber  int    `json:"card_number"`
+	CardNumber  string `json:"card_number"`
 	ExpiryMonth int    `json:"expiry_month"`
 	ExpiryYear  int    `json:"expiry_year"`
 	Currency    string `json:"currency"`